+++ scripts/static/2-produce-geo-files-static.R
@@ -115,9 +115,9 @@ my_intersections <- st_intersection(my_tracts %>%
                                       # you want a projected crs when doing area
                                       # calculations. Note using crs 4326 doesn't
                                       # change results.
-                                      st_transform(102008), 
+                                      st_transform("ESRI:102008"), 
                                     my_cbsas %>% 
-                                      st_transform(102008)) %>% 
+                                      st_transform("ESRI:102008")) %>% 
                     st_transform(4326)
 
 
@@ -128,7 +128,8 @@ my_intersections <- my_intersections %>%
 # add area of tract
 my_tracts_area <- my_tracts %>% 
   transmute(GEOID, tract_area = st_area(my_tracts)) %>% 
-  st_drop_geometry()
+  st_drop_geometry() %>% 
+  as_tibble()
 
 # calculate intersection area/tract area and filter to
 # only areas where intersection is over 99.5% of the tract's
@@ -137,10 +138,10 @@ my_tracts_area <- my_tracts %>%
 tract_cbsa_ints <- my_intersections %>% 
   st_drop_geometry() %>% 
   right_join(my_tracts_area, "GEOID") %>% 
+  as_tibble() %>% 
   mutate(perc = int_area / tract_area,
          perc = as.numeric(perc)) %>% 
-  filter(perc >= .995)
-
+  filter(perc >= .7)
 
 # Check tracts are not in multiple CBSAs
 assert("tracts are in multiple CBSAs", 
@@ -198,3 +199,166 @@ cbsa_to_county = trct_cty_cbsa %>%
 jsonlite::write_json(cbsa_to_county, 'data/processed-data/CbsaToCounty2.json')
 jsonlite::write_json(county_to_cbsa, 'data/processed-data/CountyToCbsa2.json')
 
+
+#----Create tract<>PUMA croswalk --------------------------
+
+
+# Read in pumas
+my_pumas <- st_read("data/raw-data/big/pumas.geojson")
+
+# Generate tract population centroids
+x = my_tracts %>%
+    left_join(pop_centers_2010 %>%
+        mutate(GEOID = paste0(STATEFP, COUNTYFP, TRACTCE)) %>% 
+        select(GEOID, POPULATION, LATITUDE, LONGITUDE),
+        by = "GEOID"
+    )
+
+# Use population weighted centroids to set center of tracts. 25 tracts,
+#  don't exist in the 2010 Census provided population centroids, so we just
+#  calculate the areal centroids and append. Note, for 1 tract in TX, the
+#  population centroid is just  outside of the US, so we use areal centroid
+#  instead. For one tract in AL, both the population and areal centroid are outside
+#  the US (in the middle of the Pacific Ocean), so we hardcode lat/lon based on manual
+#  survey of Google Maps
+tract_centroids = x %>%
+    st_drop_geometry() %>%
+    as_tibble() %>%
+    # For one tract in Alaska, both the population weighted cenroid and the areal
+    # end up being in the Pacific Ocean. So just for that one tract, we
+    # hardcode a lat/lon centroid based on a manual survey of Google Maps.
+    mutate(
+        LATITUDE = if_else(GEOID == "02016000100", 52.318869, LATITUDE),
+        LONGITUDE = if_else(GEOID == "02016000100", -172.452039, LONGITUDE)
+    ) %>% 
+    filter(!is.na(LATITUDE)) %>%
+    # One tract in TX that has population centroid outside of US, so we filter
+    # out and use areal centroid instead.
+    filter(!GEOID %in% c("48479001717")) %>% 
+    st_as_sf(coords = c("LONGITUDE", "LATITUDE")) %>% 
+    st_set_crs(4326)
+unjoined_tract_centroids = x %>%
+    filter(GEOID %in% c("48479001717") | is.na(LATITUDE)) %>% 
+    select(-LATITUDE, -LONGITUDE) %>% 
+    st_centroid()
+
+tract_centroids = rbind(tract_centroids, unjoined_tract_centroids)
+
+
+# Spatially join tract centroids with PUMA's. This generates a 1:1 list of
+# tracts to PUMA's
+p_t_ints = st_join(
+    tract_centroids %>%
+        select(GEOID, POPULATION) %>% 
+        st_transform("ESRI:102008"), 
+    my_pumas %>%
+        select(
+        puma_geoid = GEOID10,
+        puma_name = NAMELSAD10
+        ) %>% 
+        st_transform("ESRI:102008")) %>% 
+    st_transform(4326)
+
+# Make sure that all tracts are joined to a PUMA
+assert(nrow(p_t_ints %>% filter(is.na(puma_geoid))) == 0)
+
+puma_tract_xwalk = p_t_ints %>% 
+st_drop_geometry() %>% 
+# Filter out tracts in Puerto Rico as we're not using for this project
+filter(!startsWith(puma_geoid, "72"))
+
+puma_tract_xwalk %>% write_csv("data/processed-data/puma_tract_xwalk.csv")
+
+# Old 1:many spatial join code. This turned out to be pretty hairy because there
+# were lots of LINESTRINGS/POINTS that were intersecting and there wasn't a good
+# # area cutoff to get clean intersections areas. Code takes 15 minutes to run.
+# my_puma_tract_intersections <- st_intersection(my_tracts %>%
+#                                       select(GEOID, state_fips = STATEFP, 
+#                                       county_fips = COUNTYFP) %>% 
+#                                       #change projection to Albers equal area as 
+#                                       # you want a projected crs when doing area
+#                                       # calculations. Note using crs 4326 doesn't
+#                                       # change results.
+#                                       st_transform("ESRI:102008"), 
+#                                     my_pumas %>%
+#                                       select(
+#                                         puma_geoid = GEOID10,
+#                                         puma_name = NAMELSAD10
+#                                       ) %>% 
+#                                       st_transform("ESRI:102008")) %>% 
+#                     st_transform(4326)
+
+
+# # add area of intersections
+# my_puma_tract_intersections <- my_puma_tract_intersections %>% 
+#   mutate(int_area = st_area(my_puma_tract_intersections)) 
+
+
+#   my_puma_tract_intersections %>% select(int_area) %>% summary()
+
+
+
+# # calculate intersection area/tract area and filter to
+# # only areas where intersection is over 2% of the tract's
+# # area. This is done to exclude intersections that only overlap 
+# # with the border of a CBSA. 
+# puma_tract_ints <- my_puma_tract_intersections %>% 
+#   right_join(my_tracts_area, "GEOID") %>% 
+#   mutate(perc = int_area / tract_area,
+#          perc = as.numeric(perc)) %>% 
+#     filter(perc > 0.05)
+  
+#   puma_tract_ints %>% st_write("data/raw-data/puma_tracts_ints.geojson")
+
+
+
+
+# tracts_in_multiple_pumas = x %>%
+#     count(GEOID) %>%
+#     filter(n > 1) %>%
+#     arrange(n) %>% 
+#     pull(GEOID)
+
+# ints_tracts = puma_tract_ints %>% 
+#     filter(GEOID %in% tracts_in_multiple_pumas)
+
+# ints_tracts %>%
+#     filter(puma_geoid == "0400120") %>%
+#     select(geometry) %>% 
+#     plot(
+#         col = alpha("green", 0.4),
+#         add = T
+#     )
+
+# p_t_ints %>% filter(puma_geoid == "0400120")
+# my_pumas %>% 
+#     filter(GEOID10 == "0400120") %>% 
+#     select(geometry) %>% 
+#     plot(col = alpha("red", 0.4))
+
+# p_geoids <- x %>%
+#     filter(GEOID %in% tracts_in_multiple_pumas) %>% 
+#     pull(puma_geoid) %>%
+#     unique()
+
+# int_pumas = my_pumas %>%
+#     filter(GEOID10 %in% p_geoids)
+    
+# mapview(int_pumas)
+        
+# mapview(small_int_tracts %>% slice(1), col.regions = "blue")
+# mapview(my_pumas %>% filter())
+
+# leaflet(ints_tracts %>% slice(2)) %>%
+#     addProviderTiles('CartoDB.Positron') %>% 
+#     leafem::addFeatures()
+
+# ints_tracts %>% slice(2) %>% plot()
+# plot(ints_tracts %>% slice(3:4) %>% select(geometry), col = "green")
+# plot(int_pumas %>% filter(GEOID10 %in% (ints_tracts %>% slice(3) %>% pull(puma_geoid))) %>% select(geometry), col = "red")
+# mapview(int_pumas %>% filter(puma_geoid == ints_tracts %>% slice(1) %>% pull(puma_geoid)), col.regions = "green")
+
+
+# summary()
+# my_puma_tract_intersections %>% as_tibble()
+# my_tracts_area %>% as_tibble()