+++ scripts/update-master.R
@@ -12,25 +12,11 @@ static <- FALSE
 # set dryrun = TRUE
 dryrun <- FALSE
 
-# Are we using the BLS data for job loss calculation? Set to TRUE after
-# 2020-05-08. If FALSE, use WA/NY weighted data for job loss calculations
-is_bls <- TRUE
-
-# Number of past weeks of unemployment data to use for WA, NY AND BLS
-past_unemployment_weeks <- 7
-
 # Set bin values for legend bins and histograms. The histogram needs to be
 # manually reviewd and may need to change these values
-tmax_bins <- c(100, 150, 200, 250, 1050)
-max_bins <- c(100, 250, 500, 750, 1000, 2000, 5000, 10000, 275000)
+tmax_bins <- c(100, 150, 200, 250, 850)
+max_bins <- c(100, 250, 500, 750, 1000, 2000, 5000, 10000, 350000)
 
-
-### Below params are only for NY/WA data
-# Week number for downloading WA data. Latest week as of 2020-05-07 is 17
-week_num_wa <- 17
-
-# LODES quarter from which to start % change in employment
-start_quarter_lodes <- 3
 
 
 if (static) {
@@ -43,12 +29,10 @@ if (static) {
 }
 
 source("scripts\\update\\1-download-data-update.R", encoding = "UTF-8")
-source("scripts\\update\\1b-generate-bls-state-claims-csv-update.R", encoding = "UTF-8")
-source("scripts\\update\\2a-job-loss-by-industry-wa-update.R", encoding = "UTF-8")
-source("scripts\\update\\2b-job-loss-by-industry-bls-update.R", encoding = "UTF-8")
-source("scripts\\update\\2c-job-loss-by-industry-ny-update.R", encoding = "UTF-8")
-source("scripts\\update\\2y-job-loss-by-industry-combine-states-update.R", encoding = "UTF-8")
-source("scripts\\update\\2z-job-loss-by-industry-by-state-update.R", encoding = "UTF-8")
+source("scripts\\update\\2v4a-job-loss-projected-forward-ces.R", encoding = "UTF-8")
+source("scripts\\update\\2v4b-job-loss-by-industry-ces-sae-ipums-update.R", encoding = "UTF-8")
+source("scripts\\update\\2v4c-job-loss-by-industry-ipums-update.R", encoding = "UTF-8")
+source("scripts\\update\\2v4d-job-loss-by-industry-ipums-summary-update.R", encoding = "UTF-8")
 source("scripts\\update\\3-produce-data-files-update.R", encoding = "UTF-8")
 source("scripts\\update\\4-produce-summary-stats-update.R", encoding = "UTF-8")